refactor(frontend): extract CultivationTypeEditCell into its own module
Move the memoized CultivationTypeEditCell (and its props interface) out of
the large PlantingPlans page file into src/pages/CultivationTypeEditCell.tsx,
keeping its custom memo comparator. The now-unused memo, TextField,
MenuItem, GridRenderEditCellParams, useClosedSelectTypeahead, and
CultivationTypeSelectOption imports are dropped from the page. No behavior
change.
